requirements.txt
appnope==0.1.0
args==0.1.0
backcall==0.1.0
beautifulsoup4==4.7.1
bs4==0.0.1
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
clint==0.5.1
decorator==4.4.0
google-api-core==1.8.2
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==0.29.1
google-cloud-translate==1.3.3
googleapis-common-protos==1.5.9
httplib2==0.12.1
idna==2.5
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
jupyter-client==5.2.4
jupyter-core==4.4.0
oauth2client==4.1.3
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
protobuf==3.4.0
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
Pygments==2.3.1
python-dateutil==2.8.0
pytz==2018.9
pyzmq==18.0.1
requests==2.18.1
rsa==4.0
six==1.12.0
soupsieve==1.9
tornado==6.0.2
tqdm==4.31.1
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.21.1
wcwidth==0.1.7
